Add linting tests based on girder test framework; Rename all jsx file to js
Core girder cmake add eslint function is lacking ability of adding eslint command parameter, making eslint ignore all jsx file when running linting test. It's OK and known practice to just use js extension for jsx file, so renaming all jsx to js now to make the test work.
