+++ .gitignore
@@ -10,4 +10,3 @@ src
 /*.egg-info/
 /build/
 /dist/
-ansible/site.retry
