+++ package.json
@@ -4,14 +4,15 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "babel-preset-react": "^6.24.1",
+    "bootstrap-slider": "^9.9.0",
+    "eslint-plugin-react": "^7.4.0",
     "geojs": "0.12.1",
     "hammerjs": "^2.0.8",
-    "bootstrap-slider": "^9.9.0",
-    "babel-preset-react": "^6.24.1",
-    "react-bootstrap-slider": "2.1.3",
+    "react": "^16.0.0",
     "react-bootstrap": "^0.31.5",
+    "react-bootstrap-slider": "2.1.3",
+    "react-dom": "^16.0.0",
     "yamljs": "0.3.0"
   },
   "author": "",
