+++ web_external/App.js
@@ -2,7 +2,7 @@ import App from 'girder/views/App';
 import ReactDOM from 'react-dom';
 import React from 'react';
 
-import AppContainer from './AppContainer.jsx';
+import AppContainer from './AppContainer';
 
 import template from './templates/layout.pug';
 import './stylesheets/layout.styl';
