+++ web_external/InfoView/index.js
