+++ web_external/Viewer/VideoWidgetWrapper/index.js
