+++ webpack.helper.js
@@ -1,14 +1,11 @@
 module.exports = (config) => {
     config.module.rules.push({
-        test: /\.jsx?$/,
+        test: /\.js?$/,
         exclude: /node_modules/,
         loader: "babel-loader",
         query: {
             presets: ['es2015', 'react']
         }
     });
-    config.resolve = {
-        extensions: ['.js', '.jsx']
-    }
     return config;
 };
